drf-api/requirements.txt
django
Pillow
djangorestframework
django-heroku
